[Bronze V] Title: Multiply, Time: 112 ms, Memory: 12976 KB, Score: 40 point -BaekjoonHub

+++ 백준/Bronze/22193. Multiply/Multiply.java
@@ -6,9 +6,9 @@ public class Main {
     	
     	String temp = scan.nextLine();
     	
-    	int A = scan.nextInt();
-    	int B = scan.nextInt();
-    	int res = A * B;
+    	long A = scan.nextLong();
+    	long B = scan.nextLong();
+    	long res = A * B;
     	
     	System.out.println(res);
